Consolidate some is*Constant logic
